+++ src/main/java/StreakTheSpire/Config/CriteriaModPanelPage.java
@@ -8,22 +8,33 @@ import StreakTheSpire.StreakTheSpire;
 import StreakTheSpire.Utils.FixedModLabel;
 import StreakTheSpire.Utils.FixedModLabeledToggleButton;
 import StreakTheSpire.Utils.LocalizationConstants;
-import basemod.IUIElement;
-import basemod.ModPanel;
+import StreakTheSpire.Utils.TextureCache;
+import basemod.*;
+import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.Vector2;
-import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.helpers.FontHelper;
+import com.megacrit.cardcrawl.helpers.ImageMaster;
 import com.megacrit.cardcrawl.localization.UIStrings;
 
 import java.util.HashMap;
 
 public class CriteriaModPanelPage extends ConfigModPanelPage {
-    private boolean isInitialised = false;
 
+    private static final String minButtonNormalPath = "StreakTheSpire/textures/ui/0_button_normal.png";
+    private static final String minButtonHoveredPath = "StreakTheSpire/textures/ui/0_button_hovered.png";
+    private static final String minButtonPressedPath = "StreakTheSpire/textures/ui/0_button_pressed.png";
+    private static final String maxButtonNormalPath = "StreakTheSpire/textures/ui/20_button_normal.png";
+    private static final String maxButtonHoveredPath = "StreakTheSpire/textures/ui/20_button_hovered.png";
+    private static final String maxButtonPressedPath = "StreakTheSpire/textures/ui/20_button_pressed.png";
+
+    private boolean isInitialised = false;
     private HashMap<DisplayPreferencesModel.RenderLayer, FixedModLabeledToggleButton> renderLayerToToggleButton = new HashMap<>();
 
     @Override
-    public void initialise(ModPanel modPanel) {
+    public String getTitleLocalizationID() { return LocalizationConstants.Config.CriteriaSettingsTitle; }
+
+    @Override
+    public void initialise(ModPanel modPanel, Vector2 contentTopLeft, Vector2 contentDimensions) {
         if(isInitialised)
             return;
 
@@ -32,19 +43,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         UIStrings uiStrings = StreakTheSpire.get().getConfigUIStrings();
         StreakCriteriaModel criteriaModel = StreakTheSpire.get().getStreakCriteriaModel();
 
-        Vector2 elementPosition = ConfigModPanel.PageTopLeft.cpy();
-
-        FixedModLabel title = new FixedModLabel(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.CriteriaSettingsTitle),
-                elementPosition.x,
-                elementPosition.y,
-                modPanel,
-                (label) -> {}
-        );
-        addElement(title);
-
-
-        elementPosition.y -= ConfigModPanel.TitleLineHeight;
+        Vector2 elementPosition = contentTopLeft.cpy();
         IUIElement requireHeartKillButton = createBooleanPropElement(
                 elementPosition.x,
                 elementPosition.y,
@@ -56,17 +55,28 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(requireHeartKillButton);
 
-        //TODO: Create an encapsulated int prop element or turn it into carousel, either way, support tooltips
+        TextureCache textureCache = StreakTheSpire.get().getTextureCache();
+        Texture minNormalTex = textureCache.getTexture(minButtonNormalPath);
+        Texture minHoveredTex = textureCache.getTexture(minButtonHoveredPath);
+        Texture minPressedTex = textureCache.getTexture(minButtonPressedPath);
+        Texture maxNormalTex = textureCache.getTexture(maxButtonNormalPath);
+        Texture maxHoveredTex = textureCache.getTexture(maxButtonHoveredPath);
+        Texture maxPressedTex = textureCache.getTexture(maxButtonPressedPath);
+
         elementPosition.y -= ConfigModPanel.LineHeight;
-        IUIElement[] requiredAscensionInputElements = createIntPropElements(
+        IUIElement[] requiredAscensionInputElements = createIntPropButtonsElements(
                 elementPosition.x,
                 elementPosition.y,
-                50.0f,
-                ConfigModPanel.LineHeight * 0.8f,
                 criteriaModel.requiredAscensionLevel,
                 0,
                 20,
                 uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RequiredAscensionLevelLabel),
+                minNormalTex,
+                minHoveredTex,
+                minPressedTex,
+                maxNormalTex,
+                maxHoveredTex,
+                maxPressedTex,
                 modPanel,
                 (val) -> recalculateStreaks()
         );
@@ -74,18 +84,6 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
             addElement(element);
 
         elementPosition.y -= ConfigModPanel.LineHeight;
-        IUIElement trackContinuousButton = createBooleanPropElement(
-                elementPosition.x,
-                elementPosition.y,
-                criteriaModel.trackContinuous,
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.TrackContinuousLabel),
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.TrackContinuousTooltip),
-                modPanel,
-                (val) -> recalculateStreaks()
-        );
-        addElement(trackContinuousButton);
-
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
         IUIElement enforceRotatingButton = createBooleanPropElement(
                 elementPosition.x,
                 elementPosition.y,
@@ -97,7 +95,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(enforceRotatingButton);
 
-        elementPosition.x = ConfigModPanel.PageTopLeft.x;
+        elementPosition.x = contentTopLeft.x;
         elementPosition.y -= ConfigModPanel.TitleLineHeight;
         FixedModLabel allowTitle = new FixedModLabel(
                 uiStrings.TEXT_DICT.get(LocalizationConstants.Config.AllowTitle),
@@ -121,7 +119,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(allowCustomSeedsButton);
 
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
+        elementPosition.x += (contentDimensions.x / NumCheckboxenPerLine);
         IUIElement allowDailiesButton = createBooleanPropElement(
                 elementPosition.x,
                 elementPosition.y,
@@ -133,7 +131,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(allowDailiesButton);
 
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
+        elementPosition.x += (contentDimensions.x / NumCheckboxenPerLine);
         IUIElement allowEndlessButton = createBooleanPropElement(
                 elementPosition.x,
                 elementPosition.y,
@@ -145,7 +143,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(allowEndlessButton);
 
-        elementPosition.x = ConfigModPanel.PageTopLeft.x;
+        elementPosition.x = contentTopLeft.x;
         elementPosition.y -= ConfigModPanel.LineHeight;
         IUIElement allowBetaButton = createBooleanPropElement(
                 elementPosition.x,
@@ -158,7 +156,7 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(allowBetaButton);
 
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
+        elementPosition.x += (contentDimensions.x / NumCheckboxenPerLine);
         IUIElement allowDemoButton = createBooleanPropElement(
                 elementPosition.x,
                 elementPosition.y,
@@ -170,118 +168,6 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         );
         addElement(allowDemoButton);
 
-        // TODO: Move this to DisplayPreferencesModPanelPage if we run out of space, also implementing paging properly
-        elementPosition.x = ConfigModPanel.PageTopLeft.x;
-        elementPosition.y -= ConfigModPanel.TitleLineHeight;
-        FixedModLabel displayPreferencesTitle = new FixedModLabel(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.DisplayPreferencesTitle),
-                elementPosition.x,
-                elementPosition.y,
-                modPanel,
-                (label) -> {}
-        );
-        addElement(displayPreferencesTitle);
-
-        elementPosition.y -= ConfigModPanel.TitleLineHeight;
-        FixedModLabel renderLayerTitle = new FixedModLabel(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerTitle),
-                elementPosition.x,
-                elementPosition.y,
-                FontHelper.tipBodyFont,
-                modPanel,
-                (label) -> {}
-        );
-        addElement(renderLayerTitle);
-
-        DisplayPreferencesModel displayPreferencesModel = StreakTheSpire.get().getDisplayPreferencesModel();
-
-        elementPosition.y -= ConfigModPanel.LineHeight;
-        FixedModLabeledToggleButton button = new FixedModLabeledToggleButton(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerPreRoomLabel),
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerPreRoomTooltip),
-                elementPosition.x,
-                elementPosition.y,
-                Settings.CREAM_COLOR,
-                FontHelper.charDescFont,
-                displayPreferencesModel.renderLayer.get() == DisplayPreferencesModel.RenderLayer.PreRoom,
-                modPanel,
-                (label) -> {},
-                (toggleButton) -> {
-                    if(toggleButton.enabled)
-                        changeRenderLayer(DisplayPreferencesModel.RenderLayer.PreRoom);
-                }
-        );
-        renderLayerToToggleButton.put(DisplayPreferencesModel.RenderLayer.PreRoom, button);
-        addElement(button);
-
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
-        button = new FixedModLabeledToggleButton(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerTopPanelLabel),
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerTopPanelTooltip),
-                elementPosition.x,
-                elementPosition.y,
-                Settings.CREAM_COLOR,
-                FontHelper.charDescFont,
-                displayPreferencesModel.renderLayer.get() == DisplayPreferencesModel.RenderLayer.TopPanel,
-                modPanel,
-                (label) -> {},
-                (toggleButton) -> {
-                    if(toggleButton.enabled)
-                        changeRenderLayer(DisplayPreferencesModel.RenderLayer.TopPanel);
-                }
-        );
-        renderLayerToToggleButton.put(DisplayPreferencesModel.RenderLayer.TopPanel, button);
-        addElement(button);
-
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
-        button = new FixedModLabeledToggleButton(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerAboveMostLabel),
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerAboveMostTooltip),
-                elementPosition.x,
-                elementPosition.y,
-                Settings.CREAM_COLOR,
-                FontHelper.charDescFont,
-                displayPreferencesModel.renderLayer.get() == DisplayPreferencesModel.RenderLayer.AboveMost,
-                modPanel,
-                (label) -> {},
-                (toggleButton) -> {
-                    if(toggleButton.enabled)
-                        changeRenderLayer(DisplayPreferencesModel.RenderLayer.AboveMost);
-                }
-        );
-        renderLayerToToggleButton.put(DisplayPreferencesModel.RenderLayer.AboveMost, button);
-        addElement(button);
-
-        elementPosition.x += (ConfigModPanel.PageDimensions.x / NumCheckboxenPerLine);
-        button = new FixedModLabeledToggleButton(
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerAboveAllLabel),
-                uiStrings.TEXT_DICT.get(LocalizationConstants.Config.RenderLayerAboveAllTooltip),
-                elementPosition.x,
-                elementPosition.y,
-                Settings.CREAM_COLOR,
-                FontHelper.charDescFont,
-                displayPreferencesModel.renderLayer.get() == DisplayPreferencesModel.RenderLayer.AboveAll,
-                modPanel,
-                (label) -> {},
-                (toggleButton) -> {
-                    if(toggleButton.enabled)
-                        changeRenderLayer(DisplayPreferencesModel.RenderLayer.AboveAll);
-                }
-        );
-        renderLayerToToggleButton.put(DisplayPreferencesModel.RenderLayer.AboveAll, button);
-        addElement(button);
-
-        elementPosition.x = ConfigModPanel.PageTopLeft.x;
-        elementPosition.y = ConfigModPanel.PageTopLeft.y - ConfigModPanel.PageDimensions.y + 40f;
-        FixedModLabel defaultNoteLabel = new FixedModLabel(uiStrings.TEXT_DICT.get(LocalizationConstants.Config.DefaultNoteLabel),
-                elementPosition.x,
-                elementPosition.y,
-                FontHelper.tipBodyFont,
-                modPanel,
-                (label) -> {}
-        );
-        addElement(defaultNoteLabel);
-
         isInitialised = true;
     }
 
@@ -292,17 +178,6 @@ public class CriteriaModPanelPage extends ConfigModPanelPage {
         PlayerStreakStoreModel model = StreakTheSpire.get().getStreakStoreDataModel();
         PlayerStreakStoreController controller = new PlayerStreakStoreController(model);
         controller.calculateStreakData(criteria, true);
-
-        StreakTheSpire.get().saveConfig();
-    }
-
-    private void changeRenderLayer(DisplayPreferencesModel.RenderLayer renderLayer) {
-        DisplayPreferencesModel displayPreferencesModel = StreakTheSpire.get().getDisplayPreferencesModel();
-        displayPreferencesModel.renderLayer.set(renderLayer);
-
-        renderLayerToToggleButton.forEach((layer, button) -> {
-            button.toggle.enabled = layer == renderLayer;
-        });
 
         StreakTheSpire.get().saveConfig();
     }