+++ src/main/java/StreakTheSpire/Data/RotatingConstants.java
@@ -1,0 +1,6 @@
+package StreakTheSpire.Data;
+
+public class RotatingConstants {
+    public static String Identifier = "ROTATING";
+    public static String LocalisationID = "Rotating";
+}