+++ src/main/java/StreakTheSpire/Models/CharacterCoreDataSetModel.java
@@ -1,0 +1,9 @@
+package StreakTheSpire.Models;
+
+import StreakTheSpire.Utils.Properties.PropertyHashSet;
+
+public class CharacterCoreDataSetModel {
+    public PropertyHashSet<CharacterCoreDataModel> characterLocalisations = new PropertyHashSet<>();
+
+    public CharacterCoreDataSetModel() {}
+}